requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Migrate
Python-dotenv
Werkzeug